web-application/requirements.txt
Flask==2.0.1
requests==2.26.0
Werkzeug==2.0.1
Pillow>=10.0.0
python-dotenv>=0.19.0
gunicorn==20.1.0
tensorflow>=2.8.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-image>=0.19.0
scipy>=1.7.0
aiohttp>=3.8.0
asyncio
dataclasses
matplotlib>=3.5.0
pandas>=1.3.0